First commit, tested, of an infrastructure for version control. Additionally, smbtad is storing it's complete configuraten in a table.
For the version, smbtad is storing two values in the table, one
is the
"database version"
and another one is the
"smbtad version"

smbtad will refuse to start if the database version is not the
same as the smbtad version. It will then advice the user to
run smbtaquery -C, or smbtad -T,  and exit.

Also a module-list table is being created but this feature is not
yet implemented (listing module protocol subversions).